added Butterworth Bandpass filter, Normalization & Preemphasis filter
Updated all data inputs to be of type Float32 and improved audio processing

The commit includes several changes in alignment.py, asr.py, audio.py, requirements.txt, and transcribe.py, primarily to enforce all data inputs to be of Float32 type for better compatibility with models and libraries used. In addition, improvements have been made to audio data processing through the inclusion of preemphasis, normalization and Butterworth Bandpass filters in the load_audio function in audio.py for clearer and more precise audio waveforms. Modified functions and in alignments, asr and transcribe.py were then updated to include these changes. Lastly, new versions of the dependencies were defined in requirements.txt for compatibility purposes.

+++ whisperx/alignment.py
@@ -104,11 +104,16 @@ def align(
     """
     Align phoneme recognition predictions to known transcription.
     """
-    
+
+    model = model.to(dtype=torch.float32)
+
     if not torch.is_tensor(audio):
         if isinstance(audio, str):
             audio = load_audio(audio)
         audio = torch.from_numpy(audio)
+
+    audio = audio.to(dtype=torch.float32)
+
     if len(audio.shape) == 1:
         audio = audio.unsqueeze(0)
